+++ addon/CHANGELOG.md
@@ -1,4 +1,4 @@
-##### *GENERATED FILE [2026-07-10 20:16 UTC] — do not edit by hand.*
+##### *GENERATED FILE [2026-07-10 21:52 UTC] — do not edit by hand.*
 
 # Changelog